feat: move QR scanner into link input field and remove from settings (#291)
